Use original arm description
Since the hardware is used with multiple modifications there exist many
attached devices that are represented in the description.
It is easier to just append the pusher to the original arm description
than copying it each time there is a change.
The tams_ur5_setup.urdf.xacro now attaches the pusher link.