Replace endeffector pusher description with attached collision object
Loading the pusher from urdf allows using it as an endeffector.
However this means that the launch dependencies and configuration
need to be copied what leads to possibly high maintenance.
This implementation extends the move group interface and allows
dynamic attachement of pushers.
Additional functions offer the move group functionality as if the
pusher is the endeffector.